Heavy auth clean up
