+++ api/src/integrations/index.ts
@@ -1,1 +1,2 @@
+export { auth0Integration } from "./auth0-integration"
 export { yukonGovernmentIntegration } from "./yukon-government-integration"
