+++ api/src/routes/index.ts
@@ -3,7 +3,7 @@ import { DatabaseError } from "sequelize"
 
 import logger from "@/utils/logger"
 import { GIT_COMMIT_HASH, RELEASE_TAG } from "@/config"
-import { databaseHealthCheckMiddleware, checkJwt, loadUser } from "@/middleware"
+import { databaseHealthCheckMiddleware, checkJwt, authorizationMiddleware } from "@/middleware"
 import { healthCheckRouter } from "@/routes/healthcheck-router"
 import {
   Expenses,
@@ -61,7 +61,7 @@ router.use("/api/lookup-tables", databaseHealthCheckMiddleware, lookupTableRoute
 //// END LEGACY ROUTES
 
 // api routes
-router.use("/api", databaseHealthCheckMiddleware, checkJwt, loadUser)
+router.use("/api", databaseHealthCheckMiddleware, checkJwt, authorizationMiddleware)
 
 //// START MORE LEGACY ROUTES
 router.use("/api/form", formRouter)
